moved wdbErrors to pkg/wdb
