+++ internal/namespaces/namespaces.go
@@ -3,8 +3,8 @@ package namespaces
 import (
 	"fmt"
 
-	er "github.com/TanmoySG/wunderDB/internal/errors"
 	"github.com/TanmoySG/wunderDB/model"
+	er "github.com/TanmoySG/wunderDB/pkg/wdb/errors"
 )
 
 type Namespaces map[model.Identifier]*model.Namespace
